Function eiffelvis_core::algorithms::depth_first

pub fn depth_first<G, F>(graph: &G, index: G::Idx, callback: &mut F) where
    G: Graph,
    F: FnMut(G::Idx) -> ControlFlow<()>, 

A recursive depth first traversal implementation, if user provided closure returns ControlFlow::Break that particular node’s edges will not be visited, but will not stop the whole traversal.