Module eiffelvis_core::domain::user_queries

Query types and functions to acquire specific data from an [EiffelGraph]

Structs

EventFilterMeta

Query

Describes a query to collect nodes from an eiffel graph, to match an event all filters need to match.

TrackedQuery

A tracked query, only returns new matches.

Enums

Collection

Used collection method, selected variant is run after filtering has been done, this means you can get nodes back that may not filfull the filter requirements.

EventFilter

RangeFilterBound