Struct eiffelvis_core::graph_storage::chunked_storage::ChunkedGraph

pub struct ChunkedGraph<K: Key, N, E> { /* fields omitted */ }

Implementations

impl<K: Key, N, E> ChunkedGraph<K, N, E>

pub fn new(max_chunks: usize, chunk_size: u32) -> Self

pub fn to_index(&self, key: K) -> Option<ChunkedIndex>

Translates given key to internal index. This is generally an relatively expensive operation, avoid frequent translations.

pub fn from_index(&self, index: ChunkedIndex) -> Option<K>

Translates given internal index to key

pub fn chunks(&self) -> usize

pub fn last(&self) -> Option<ChunkedIndex>

Returns the index of the most recent added node

pub fn node_count(&self) -> usize

pub fn cmp_index(&self, lhs: ChunkedIndex, rhs: ChunkedIndex) -> Ordering